refactor(reducer): modularize diagnostic parsers
