+++ crates/bazel-mcp-reducer/src/build.rs
@@ -3,12 +3,13 @@ use bazel_mcp_bep::{
     view::{BuildEventIdView, FileView, NamedSetOfFilesView, build_event, build_event_id, file},
 };
 use bazel_mcp_types::{
-    Artifact, ArtifactKind, Diagnostic, DiagnosticCategory, DiagnosticLocation, InvocationSummary,
-    Severity, TargetCounts, TargetResult, TestCounts, TestResult, TestStatus,
+    Artifact, ArtifactKind, Diagnostic, DiagnosticCategory, InvocationSummary, Severity,
+    TargetCounts, TargetResult, TestCounts, TestResult, TestStatus,
 };
 use std::collections::{BTreeMap, BTreeSet};
 
-use crate::{Budget, ReducerEvent, ReducerEventKind, deduplicate_lines, normalize_terminal_text};
+use crate::diagnostics::{add_text_diagnostics, bounded_text};
+use crate::{Budget, ReducerEvent, ReducerEventKind};
 
 pub struct ReductionInput<'a> {
     pub events: &'a [BepEvent],
@@ -851,1561 +852,6 @@ fn file_artifact(file: &FileView<'_>) -> Option<Artifact> {
     })
 }
 
-fn bounded_text(value: &str, maximum_bytes: usize) -> String {
-    if value.len() <= maximum_bytes {
-        return value.to_owned();
-    }
-    let mut boundary = maximum_bytes;
-    while !value.is_char_boundary(boundary) {
-        boundary -= 1;
-    }
-    format!("{}…", &value[..boundary])
-}
-
-fn add_text_diagnostics(input: &[u8], diagnostics: &mut Vec<Diagnostic>) {
-    let normalized = normalize_terminal_text(input);
-    let swc = parse_swc_diagnostics(&normalized);
-    let has_swc_diagnostics = !swc.diagnostics.is_empty();
-    diagnostics.extend(swc.diagnostics);
-    let mut cpp_linker_parser = CppLinkerDiagnosticParser::default();
-    for line in normalized.lines() {
-        if let Some(diagnostic) = cpp_linker_parser.observe_line(line) {
-            diagnostics.push(diagnostic);
-        }
-    }
-    diagnostics.extend(parse_java_compiler_diagnostics(&normalized));
-    let mut javascript_parser = JavaScriptTestDiagnosticParser::default();
-    let mut javascript_test_messages = BTreeSet::new();
-    for line in normalized.lines() {
-        if let Some(diagnostic) = javascript_parser.observe_line(line) {
-            javascript_test_messages.insert(diagnostic.message.clone());
-            diagnostics.push(diagnostic);
-        }
-    }
-    if let Some(diagnostic) = javascript_parser.finish() {
-        javascript_test_messages.insert(diagnostic.message.clone());
-        diagnostics.push(diagnostic);
-    }
-    let mut java_parser = JavaTestDiagnosticParser::default();
-    let mut java_test_messages = BTreeSet::new();
-    for line in normalized.lines() {
-        if let Some(diagnostic) = java_parser.observe_line(line) {
-            java_test_messages.insert(diagnostic.message.clone());
-            diagnostics.push(diagnostic);
-        }
-    }
-    if let Some(diagnostic) = java_parser.finish() {
-        java_test_messages.insert(diagnostic.message.clone());
-        diagnostics.push(diagnostic);
-    }
-    let mut starlark_parser = StarlarkDiagnosticParser::default();
-    for line in normalized.lines() {
-        if let Some(diagnostic) = starlark_parser.observe_line(line) {
-            diagnostics.push(diagnostic);
-        }
-    }
-    let mut python_parser = PythonDiagnosticParser::default();
-    for line in normalized.lines() {
-        if javascript_exception_message(line)
-            .is_some_and(|message| javascript_test_messages.contains(message))
-            || java_exception_message(line)
-                .is_some_and(|message| java_test_messages.contains(message))
-        {
-            continue;
-        }
-        if let Some(diagnostic) = python_parser.observe_line(line) {
-            diagnostics.push(diagnostic);
-        }
-    }
-    let candidates = deduplicate_lines(&normalized);
-    let has_strict_dependency_block = candidates.iter().any(|(line, _)| {
-        line.to_ascii_lowercase()
-            .contains("missing strict dependencies")
-    });
-    let strict_dependency_count = candidates
-        .iter()
-        .filter(|(line, _)| strict_dependency_diagnostic(line).is_some())
-        .count();
-
-    for (line, count) in candidates {
-        if swc.consumed_lines.contains(line.trim())
-            || (has_swc_diagnostics && is_swc_action_wrapper(&line))
-        {
-            continue;
-        }
-        if has_strict_dependency_block
-            && let Some(mut diagnostic) = strict_dependency_diagnostic(&line)
-        {
-            diagnostic.repetition_count = count;
-            diagnostics.push(diagnostic);
-            continue;
-        }
-        if let Some(mut diagnostic) = parse_cpp_diagnostic(&line) {
-            diagnostic.repetition_count = count;
-            diagnostics.push(diagnostic);
-            continue;
-        }
-        if let Some(mut diagnostic) = parse_typescript_diagnostic(&line) {
-            diagnostic.repetition_count = count;
-            diagnostics.push(diagnostic);
-            continue;
-        }
-        if let Some(mut diagnostic) = parse_protobuf_diagnostic(&line) {
-            diagnostic.repetition_count = count;
-            diagnostics.push(diagnostic);
-            continue;
-        }
-        if let Some(mut diagnostic) = parse_go_diagnostic(&line) {
-            diagnostic.repetition_count = count;
-            diagnostics.push(diagnostic);
-            continue;
-        }
-        if parse_cpp_linker_diagnostic(&line).is_some() {
-            continue;
-        }
-        if parse_java_compiler_diagnostic(&line).is_some()
-            || javascript_exception_message(&line)
-                .is_some_and(|message| javascript_test_messages.contains(message))
-            || java_exception_message(&line)
-                .is_some_and(|message| java_test_messages.contains(message))
-        {
-            continue;
-        }
-        if parse_starlark_inline_diagnostic(&line)
-            .is_some_and(|diagnostic| is_starlark_root_cause_message(&diagnostic.message))
-            || starlark_error_message(&line).is_some()
-            || is_starlark_traceback_header(&line)
-        {
-            continue;
-        }
-        if parse_python_location(&line).is_some() || python_exception_message(&line).is_some() {
-            continue;
-        }
-        if has_strict_dependency_block
-            && strict_dependency_count == 0
-            && line
-                .to_ascii_lowercase()
-                .contains("missing strict dependencies")
-        {
-            diagnostics.push(Diagnostic {
-                severity: Severity::Error,
-                category: DiagnosticCategory::Compilation,
-                message: line,
-                location: None,
-                target: None,
-                action: None,
-                repetition_count: count,
-            });
-            continue;
-        }
-        if !is_actionable(&line) {
-            continue;
-        }
-        diagnostics.push(Diagnostic {
-            severity: if line.to_ascii_lowercase().contains("warning:") {
-                Severity::Warning
-            } else {
-                Severity::Error
-            },
-            category: category_from_text(&line),
-            message: line,
-            location: None,
-            target: None,
-            action: None,
-            repetition_count: count,
-        });
-    }
-}
-
-fn parse_cpp_diagnostic(line: &str) -> Option<Diagnostic> {
-    let line = line
-        .trim()
-        .strip_prefix("ERROR: ")
-        .unwrap_or_else(|| line.trim());
-    let (path, line_number, column, message) =
-        parse_cpp_colon_location(line).or_else(|| parse_cpp_parenthesized_location(line))?;
-    if path.contains(": ") {
-        return None;
-    }
-    let (severity, message) = parse_cpp_severity_message(message)?;
-    Some(Diagnostic {
-        severity,
-        category: DiagnosticCategory::Compilation,
-        message: message.to_owned(),
-        location: Some(DiagnosticLocation {
-            path: compact_cpp_path(path),
-            line: Some(line_number),
-            column,
-        }),
-        target: None,
-        action: None,
-        repetition_count: 1,
-    })
-}
-
-fn parse_cpp_colon_location(line: &str) -> Option<(&str, u32, Option<u32>, &str)> {
-    let path_end = cpp_path_end(line, ':')?;
-    let (line_number, remainder) = split_u32_prefix(&line[path_end + 1..])?;
-    let (column, message) = split_u32_prefix(remainder)
-        .map_or((None, remainder), |(column, message)| {
-            (Some(column), message)
-        });
-    Some((&line[..path_end], line_number, column, message))
-}
-
-fn parse_cpp_parenthesized_location(line: &str) -> Option<(&str, u32, Option<u32>, &str)> {
-    let path_end = cpp_path_end(line, '(')?;
-    let remainder = line[path_end..].strip_prefix('(')?;
-    let (coordinates, message) = remainder
-        .split_once("): ")
-        .or_else(|| remainder.split_once("):"))?;
-    let (line_number, column) = coordinates.split_once(',').map_or_else(
-        || (coordinates.trim().parse::<u32>().ok(), None),
-        |(line_number, column)| {
-            (
-                line_number.trim().parse::<u32>().ok(),
-                column.trim().parse::<u32>().ok(),
-            )
-        },
-    );
-    Some((&line[..path_end], line_number?, column, message))
-}
-
-fn parse_cpp_severity_message(message: &str) -> Option<(Severity, &str)> {
-    let message = message.trim();
-    for (marker, severity) in [
-        ("fatal error", Severity::Error),
-        ("error", Severity::Error),
-        ("warning", Severity::Warning),
-        ("note", Severity::Note),
-    ] {
-        let Some(remainder) = message.strip_prefix(marker) else {
-            continue;
-        };
-        let remainder = remainder
-            .strip_prefix(':')
-            .or_else(|| remainder.strip_prefix(' '))?
-            .trim();
-        if !remainder.is_empty() {
-            return Some((severity, remainder));
-        }
-    }
-    None
-}
-
-fn cpp_path_end(line: &str, delimiter: char) -> Option<usize> {
-    const EXTENSIONS: [&str; 19] = [
-        ".cpp", ".cxx", ".c++", ".cc", ".hpp", ".hxx", ".h++", ".hh", ".ipp", ".tpp", ".inc",
-        ".cuh", ".cu", ".mm", ".m", ".C", ".H", ".c", ".h",
-    ];
-    EXTENSIONS
-        .iter()
-        .filter_map(|extension| {
-            line.rmatch_indices(extension).find_map(|(index, _)| {
-                let path_end = index + extension.len();
-                line[path_end..].starts_with(delimiter).then_some(path_end)
-            })
-        })
-        .max()
-}
-
-#[derive(Debug, Default)]
-struct CppLinkerDiagnosticParser {
-    apple_undefined_symbols: bool,
-    symbols_seen: usize,
-}
-
-impl CppLinkerDiagnosticParser {
-    const MAX_SYMBOLS: usize = 64;
-
-    fn observe_line(&mut self, line: &str) -> Option<Diagnostic> {
-        let line = line.trim();
-        if line.starts_with("Undefined symbols for architecture ") {
-            self.apple_undefined_symbols = true;
-            return None;
-        }
-        if self.apple_undefined_symbols {
-            if let Some(symbol) = parse_apple_undefined_symbol(line) {
-                if self.symbols_seen >= Self::MAX_SYMBOLS {
-                    return None;
-                }
-                self.symbols_seen = self.symbols_seen.saturating_add(1);
-                return Some(cpp_linker_diagnostic(
-                    format!("undefined symbol: {}", bounded_text(symbol, 1_000)),
-                    None,
-                ));
-            }
-            if line.starts_with("ld:") || line.starts_with("clang:") {
-                self.apple_undefined_symbols = false;
-            }
-        }
-        parse_cpp_linker_diagnostic(line)
-    }
-}
-
-fn parse_apple_undefined_symbol(line: &str) -> Option<&str> {
-    let symbol = line.strip_prefix('"')?;
-    let (symbol, _) = symbol.split_once("\", referenced from:")?;
-    (!symbol.is_empty()).then_some(symbol)
-}
-
-fn parse_cpp_linker_diagnostic(line: &str) -> Option<Diagnostic> {
-    let line = line.trim();
-    if let Some(index) = line.find("undefined reference to ") {
-        let symbol = trim_linker_symbol(&line[index + "undefined reference to ".len()..]);
-        if symbol.is_empty() {
-            return None;
-        }
-        return Some(cpp_linker_diagnostic(
-            format!("undefined reference to {symbol}"),
-            parse_cpp_linker_location(&line[..index]),
-        ));
-    }
-    if let Some(index) = line.find("undefined symbol:") {
-        let symbol = trim_linker_symbol(&line[index + "undefined symbol:".len()..]);
-        if symbol.is_empty() {
-            return None;
-        }
-        return Some(cpp_linker_diagnostic(
-            format!("undefined symbol: {symbol}"),
-            None,
-        ));
-    }
-    if let Some(index) = line.find("unresolved external symbol ") {
-        let remainder = &line[index + "unresolved external symbol ".len()..];
-        let symbol = remainder
-            .split_once(" referenced in ")
-            .map_or(remainder, |(symbol, _)| symbol)
-            .trim();
-        if symbol.is_empty() {
-            return None;
-        }
-        return Some(cpp_linker_diagnostic(
-            format!("unresolved external symbol {symbol}"),
-            None,
-        ));
-    }
-    None
-}
-
-fn parse_cpp_linker_location(prefix: &str) -> Option<DiagnosticLocation> {
-    let path_end = cpp_path_end(prefix, ':')?;
-    let (line_number, _) = split_u32_prefix(&prefix[path_end + 1..])?;
-    let path = prefix[..path_end]
-        .rsplit_once(": ")
-        .map_or(&prefix[..path_end], |(_, path)| path);
-    Some(DiagnosticLocation {
-        path: compact_cpp_path(path),
-        line: Some(line_number),
-        column: None,
-    })
-}
-
-fn trim_linker_symbol(symbol: &str) -> &str {
-    symbol
-        .trim()
-        .trim_end_matches(':')
-        .trim_matches(|character| matches!(character, '`' | '\'' | '"'))
-}
-
-fn cpp_linker_diagnostic(message: String, location: Option<DiagnosticLocation>) -> Diagnostic {
-    Diagnostic {
-        severity: Severity::Error,
-        category: DiagnosticCategory::Compilation,
-        message,
-        location,
-        target: None,
-        action: None,
-        repetition_count: 1,
-    }
-}
-
-fn compact_cpp_path(path: &str) -> String {
-    let path = strip_workspace_marker(path.trim_matches('"').replace('\\', "/"));
-    if let Some((_, after_execroot)) = path.rsplit_once("/execroot/")
-        && let Some((_, relative)) = after_execroot.split_once('/')
-    {
-        return relative.to_owned();
-    }
-    path.strip_prefix("./").unwrap_or(&path).to_owned()
-}
-
-#[derive(Debug, Default)]
-struct SwcParseOutput {
-    diagnostics: Vec<Diagnostic>,
-    consumed_lines: BTreeSet<String>,
-}
-
-fn parse_swc_diagnostics(input: &str) -> SwcParseOutput {
-    const MAX_HEADER_DISTANCE: usize = 3;
-    const MAX_FRAME_LINES: usize = 8;
-
-    let lines = input.lines().collect::<Vec<_>>();
-    let mut output = SwcParseOutput::default();
-    for (location_index, line) in lines.iter().enumerate() {
-        let Some(location) = parse_swc_source_frame_location(line) else {
-            continue;
-        };
-        let Some((header_index, severity, message)) = lines[..location_index]
-            .iter()
-            .enumerate()
-            .rev()
-            .take(MAX_HEADER_DISTANCE)
-            .find_map(|(index, line)| {
-                parse_swc_message_header(line).map(|(severity, message)| (index, severity, message))
-            })
-        else {
-            continue;
-        };
-
-        let mut frame_end = location_index;
-        let mut source_line = None;
-        let mut first_source_line = None;
-        let mut caret_line = None;
-        for (index, frame_line) in lines
-            .iter()
-            .enumerate()
-            .skip(location_index + 1)
-            .take(MAX_FRAME_LINES)
-        {
-            if parse_swc_message_header(frame_line).is_some()
-                || parse_swc_source_frame_location(frame_line).is_some()
-            {
-                break;
-            }
-            frame_end = index;
-            output.consumed_lines.insert(frame_line.trim().to_owned());
-            if let Some(line_number) = swc_source_line_number(frame_line) {
-                let bounded = bounded_text(frame_line.trim(), 256);
-                first_source_line.get_or_insert_with(|| bounded.clone());
-                if location.line == Some(line_number) {
-                    source_line = Some(bounded);
-                }
-            } else if caret_line.is_none() && is_swc_caret_line(frame_line) {
-                caret_line = Some(bounded_text(frame_line.trim(), 256));
-            }
-            if is_swc_frame_terminator(frame_line) {
-                break;
-            }
-        }
-
-        output
-            .consumed_lines
-            .insert(lines[header_index].trim().to_owned());
-        output
-            .consumed_lines
-            .insert(lines[location_index].trim().to_owned());
-        let (context, context_lines) = swc_failure_context(&lines, header_index, frame_end);
-        for context_line in context_lines {
-            output
-                .consumed_lines
-                .insert(lines[context_line].trim().to_owned());
-        }
-
-        let mut message = bounded_text(message, 1_024);
-        if let Some(context) = context {
-            message.push_str("\nSWC context: ");
-            message.push_str(&bounded_text(&context, 512));
-        }
-        if let Some(source_line) = source_line.or(first_source_line) {
-            message.push('\n');
-            message.push_str(&source_line);
-        }
-        if let Some(caret_line) = caret_line {
-            message.push('\n');
-            message.push_str(&caret_line);
-        }
-        output.diagnostics.push(Diagnostic {
-            severity,
-            category: DiagnosticCategory::Compilation,
-            message,
-            location: Some(location),
-            target: None,
-            action: None,
-            repetition_count: 1,
-        });
-    }
-    output
-}
-
-fn parse_swc_message_header(line: &str) -> Option<(Severity, &str)> {
-    let line = line.trim();
-    let (severity, message) = if let Some(message) = line.strip_prefix("error:") {
-        (Severity::Error, message)
-    } else if let Some(message) = line.strip_prefix("warning:") {
-        (Severity::Warning, message)
-    } else if let Some(message) = strip_swc_icon(line, 'x').or_else(|| strip_swc_icon(line, '×')) {
-        (Severity::Error, message)
-    } else {
-        (Severity::Warning, strip_swc_icon(line, '!')?)
-    };
-    let message = message.trim();
-    (!message.is_empty()).then_some((severity, message))
-}
-
-fn strip_swc_icon(line: &str, icon: char) -> Option<&str> {
-    let remainder = line.strip_prefix(icon)?;
-    remainder
-        .chars()
-        .next()
-        .is_some_and(char::is_whitespace)
-        .then(|| remainder.trim_start())
-}
-
-fn parse_swc_source_frame_location(line: &str) -> Option<DiagnosticLocation> {
-    let line = line.trim();
-    let opening = line.find('[')?;
-    let marker = line[..opening].trim_end();
-    if !marker.ends_with(",-") && !marker.ends_with("╭─") {
-        return None;
-    }
-    let closing = line[opening + 1..].find(']')? + opening + 1;
-    let coordinates = &line[opening + 1..closing];
-    let (path_and_line, column) = coordinates.rsplit_once(':')?;
-    let (path, line_number) = path_and_line.rsplit_once(':')?;
-    if path.trim().is_empty() {
-        return None;
-    }
-    Some(DiagnosticLocation {
-        path: compact_javascript_path(path.trim()),
-        line: Some(line_number.parse::<u32>().ok()?),
-        column: Some(column.parse::<u32>().ok()?),
-    })
-}
-
-fn swc_source_line_number(line: &str) -> Option<u32> {
-    let line = line.trim();
-    let (line_number, _) = line.split_once('|').or_else(|| line.split_once('│'))?;
-    let line_number = line_number.trim();
-    (!line_number.is_empty() && line_number.bytes().all(|byte| byte.is_ascii_digit()))
-        .then(|| line_number.parse::<u32>().ok())
-        .flatten()
-}
-
-fn is_swc_caret_line(line: &str) -> bool {
-    let line = line.trim();
-    line.contains('^')
-        && (line.starts_with(':')
-            || line
-                .split_once('|')
-                .is_some_and(|(prefix, _)| prefix.trim().is_empty()))
-}
-
-fn is_swc_frame_terminator(line: &str) -> bool {
-    let line = line.trim();
-    (line.starts_with('`') && line.contains("---")) || (line.starts_with('╰') && line.contains('─'))
-}
-
-fn swc_failure_context(
-    lines: &[&str],
-    header_index: usize,
-    frame_end: usize,
-) -> (Option<String>, Vec<usize>) {
-    const MAX_CONTEXT_DISTANCE: usize = 12;
-    const MAX_CONTEXT_ITEMS: usize = 2;
-
-    let mut contexts = Vec::new();
-    let mut consumed = Vec::new();
-    let preceding_start = header_index.saturating_sub(MAX_CONTEXT_DISTANCE);
-    for (index, line) in lines
-        .iter()
-        .enumerate()
-        .take(header_index)
-        .skip(preceding_start)
-    {
-        if let Some(context) = parse_swc_context_line(line)
-            && !contexts.contains(&context)
-        {
-            contexts.push(context);
-            consumed.push(index);
-        }
-        if contexts.len() >= MAX_CONTEXT_ITEMS {
-            break;
-        }
-    }
-    for (index, line) in lines
-        .iter()
-        .enumerate()
-        .skip(frame_end + 1)
-        .take(MAX_CONTEXT_DISTANCE)
-    {
-        if parse_swc_message_header(line).is_some()
-            || parse_swc_source_frame_location(line).is_some()
-        {
-            break;
-        }
-        if let Some(context) = parse_swc_context_line(line)
-            && !contexts.contains(&context)
-        {
-            contexts.push(context);
-            consumed.push(index);
-        }
-        if contexts.len() >= MAX_CONTEXT_ITEMS {
-            break;
-        }
-    }
-    contexts.truncate(MAX_CONTEXT_ITEMS);
-    (
-        (!contexts.is_empty()).then(|| contexts.join("; ")),
-        consumed,
-    )
-}
-
-fn parse_swc_context_line(line: &str) -> Option<String> {
-    if parse_swc_message_header(line).is_some() {
-        return None;
-    }
-    let mut line = line.trim();
-    if let Some(remainder) = line.strip_prefix("Caused by:") {
-        line = remainder.trim();
-    }
-    if let Some((number, remainder)) = line.split_once(": ")
-        && !number.is_empty()
-        && number.bytes().all(|byte| byte.is_ascii_digit())
-    {
-        line = remainder.trim();
-    }
-    let lower = line.to_ascii_lowercase();
-    let is_context = lower.starts_with("failed to process")
-        || lower.starts_with("failed to parse")
-        || lower.starts_with("failed to transform")
-        || lower.starts_with("failed to invoke plugin")
-        || lower.starts_with("failed to handle")
-        || (lower.contains("plugin") && lower.contains("failed"));
-    (is_context && !line.is_empty()).then(|| line.to_owned())
-}
-
-fn is_swc_action_wrapper(line: &str) -> bool {
-    let line = line.trim();
-    let lower = line.to_ascii_lowercase();
-    line.starts_with("ERROR:")
-        && (lower.contains("build did not complete successfully")
-            || lower.contains("failed: error executing")
-            || (lower.contains("swc")
-                && (lower.contains("failed:") || lower.contains("action failed"))))
-}
-
-fn parse_typescript_diagnostic(line: &str) -> Option<Diagnostic> {
-    let line = line
-        .trim()
-        .strip_prefix("ERROR: ")
-        .unwrap_or_else(|| line.trim());
-    let (path, line_number, column, message) = parse_typescript_parenthesized_location(line)
-        .or_else(|| parse_typescript_pretty_location(line))?;
-    let message = message.trim().trim_start_matches('-').trim();
-    let (severity, message) = if let Some(message) = message.strip_prefix("error ") {
-        (Severity::Error, message)
-    } else {
-        (Severity::Warning, message.strip_prefix("warning ")?)
-    };
-    let (code, message) = message.split_once(':')?;
-    if !code.strip_prefix("TS").is_some_and(|digits| {
-        !digits.is_empty() && digits.bytes().all(|byte| byte.is_ascii_digit())
-    }) {
-        return None;
-    }
-    let message = message.trim();
-    if message.is_empty() {
-        return None;
-    }
-    Some(Diagnostic {
-        severity,
-        category: DiagnosticCategory::Compilation,
-        message: format!("{code}: {message}"),
-        location: Some(DiagnosticLocation {
-            path: compact_javascript_path(path),
-            line: Some(line_number),
-            column: Some(column),
-        }),
-        target: None,
-        action: None,
-        repetition_count: 1,
-    })
-}
-
-fn parse_typescript_parenthesized_location(line: &str) -> Option<(&str, u32, u32, &str)> {
-    let path_end = typescript_path_end(line, '(')?;
-    let remainder = line[path_end..].strip_prefix('(')?;
-    let (coordinates, message) = remainder.split_once("): ")?;
-    let (line_number, column) = coordinates.split_once(',')?;
-    let line_number = line_number.trim().parse::<u32>().ok()?;
-    let column = column.trim().parse::<u32>().ok()?;
-    Some((&line[..path_end], line_number, column, message))
-}
-
-fn parse_typescript_pretty_location(line: &str) -> Option<(&str, u32, u32, &str)> {
-    let path_end = typescript_path_end(line, ':')?;
-    let (line_number, remainder) = line[path_end + 1..].split_once(':')?;
-    let line_number = line_number.parse::<u32>().ok()?;
-    let (column, message) = remainder
-        .split_once(" - ")
-        .or_else(|| remainder.split_once(':'))?;
-    let column = column.parse::<u32>().ok()?;
-    Some((&line[..path_end], line_number, column, message))
-}
-
-fn typescript_path_end(line: &str, delimiter: char) -> Option<usize> {
-    const EXTENSIONS: [&str; 8] = [".tsx", ".mts", ".cts", ".ts", ".jsx", ".mjs", ".cjs", ".js"];
-    EXTENSIONS
-        .iter()
-        .filter_map(|extension| {
-            let marker = format!("{extension}{delimiter}");
-            line.rfind(&marker).map(|index| index + extension.len())
-        })
-        .max()
-}
-
-/// Stateful extractor for Node.js exceptions and their application frames.
-#[derive(Debug, Default)]
-pub struct JavaScriptTestDiagnosticParser {
-    leading_location: Option<DiagnosticLocation>,
-    pending: Option<Diagnostic>,
-    frames_seen: usize,
-}
-
-impl JavaScriptTestDiagnosticParser {
-    const MAX_STACK_FRAMES: usize = 64;
-
-    /// Observes one normalized test-log line and emits an exception after a
-    /// JavaScript source header or application stack frame confirms it.
-    pub fn observe_line(&mut self, line: &str) -> Option<Diagnostic> {
-        if !line.trim_start().starts_with("at ")
-            && let Some(location) = parse_javascript_location(line.trim())
-        {
-            let previous = self.take_confirmed();
-            self.leading_location = Some(location);
-            return previous;
-        }
-        if let Some(message) = javascript_exception_message(line) {
-            let leading_location = self.leading_location.take();
-            let previous = self.take_confirmed();
-            self.pending = Some(Diagnostic {
-                severity: Severity::Error,
-                category: DiagnosticCategory::Test,
-                message: message.to_owned(),
-                location: leading_location,
-                target: None,
-                action: None,
-                repetition_count: 1,
-            });
-            self.frames_seen = 0;
-            return previous;
-        }
-        self.pending.as_ref()?;
-        if let Some(location) = parse_javascript_stack_frame(line) {
-            self.frames_seen = self.frames_seen.saturating_add(1);
-            if let Some(location) = location {
-                let mut diagnostic = self.pending.take()?;
-                diagnostic.location = Some(location);
-                self.frames_seen = 0;
-                return Some(diagnostic);
-            }
-            if self.frames_seen >= Self::MAX_STACK_FRAMES {
-                return self.take_confirmed();
-            }
-            return None;
-        }
-        if line.trim().is_empty() {
-            return None;
-        }
-        self.take_confirmed()
-    }
-
-    /// Emits a confirmed exception that reached end-of-file.
-    pub fn finish(&mut self) -> Option<Diagnostic> {
-        self.take_confirmed()
-    }
-
-    fn take_confirmed(&mut self) -> Option<Diagnostic> {
-        let confirmed = self
-            .pending
-            .as_ref()
-            .is_some_and(|diagnostic| diagnostic.location.is_some())
-            || self.frames_seen > 0;
-        self.frames_seen = 0;
-        self.leading_location = None;
-        if confirmed {
-            self.pending.take()
-        } else {
-            self.pending = None;
-            None
-        }
-    }
-}
-
-fn javascript_exception_message(line: &str) -> Option<&str> {
-    let line = line.trim();
-    let exception_type = line.split_once(':').map_or(line, |(name, _)| name);
-    let class_name = exception_type.split_whitespace().next()?;
-    (!class_name.is_empty()
-        && class_name
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'_' | b'.' | b'$'))
-        && (class_name.ends_with("Error") || class_name.ends_with("Exception")))
-    .then_some(line)
-}
-
-fn parse_javascript_stack_frame(line: &str) -> Option<Option<DiagnosticLocation>> {
-    let frame = line.trim().strip_prefix("at ")?;
-    let source = if let Some((_, source)) = frame.rsplit_once('(') {
-        source.strip_suffix(')')?
-    } else {
-        frame.split_whitespace().last()?
-    };
-    if source.starts_with("node:") || matches!(source, "native" | "<anonymous>") {
-        return Some(None);
-    }
-    let location = parse_javascript_location(source)?;
-    let framework = location.path.contains("/node_modules/")
-        || location.path.starts_with("node_modules/")
-        || location.path.contains("/external/");
-    Some((!framework).then_some(location))
-}
-
-fn parse_javascript_location(value: &str) -> Option<DiagnosticLocation> {
-    let value = value.trim().trim_matches('"');
-    let path_end = javascript_path_end(value)?;
-    let coordinates = value[path_end..].strip_prefix(':')?;
-    let (line_number, column) = if let Some((line_number, column)) = coordinates.split_once(':') {
-        (
-            line_number.parse::<u32>().ok()?,
-            Some(column.parse::<u32>().ok()?),
-        )
-    } else {
-        (coordinates.parse::<u32>().ok()?, None)
-    };
-    Some(DiagnosticLocation {
-        path: compact_javascript_path(&value[..path_end]),
-        line: Some(line_number),
-        column,
-    })
-}
-
-fn javascript_path_end(value: &str) -> Option<usize> {
-    const EXTENSIONS: [&str; 8] = [".tsx", ".mts", ".cts", ".ts", ".jsx", ".mjs", ".cjs", ".js"];
-    EXTENSIONS
-        .iter()
-        .filter_map(|extension| {
-            let marker = format!("{extension}:");
-            value.rfind(&marker).map(|index| index + extension.len())
-        })
-        .max()
-}
-
-fn compact_javascript_path(path: &str) -> String {
-    let path = path
-        .trim_matches('"')
-        .strip_prefix("file://")
-        .unwrap_or(path)
-        .replace('\\', "/");
-    let path = strip_workspace_marker(path);
-    for marker in [".runfiles/_main/", ".runfiles/__main__/"] {
-        if let Some((_, relative)) = path.rsplit_once(marker) {
-            return relative.to_owned();
-        }
-    }
-    if let Some((_, after_execroot)) = path.rsplit_once("/execroot/")
-        && let Some((_, relative)) = after_execroot.split_once('/')
-    {
-        return relative.to_owned();
-    }
-    path.strip_prefix("./").unwrap_or(&path).to_owned()
-}
-
-fn parse_protobuf_diagnostic(line: &str) -> Option<Diagnostic> {
-    let marker = line.rfind(".proto:")?;
-    let path_end = marker + ".proto".len();
-    let path = line[..path_end]
-        .trim()
-        .strip_prefix("ERROR: ")
-        .unwrap_or_else(|| line[..path_end].trim());
-    let (line_number, remainder) = split_u32_prefix(&line[path_end + 1..])?;
-    let (column, message) = split_u32_prefix(remainder)
-        .map_or((None, remainder), |(column, message)| {
-            (Some(column), message)
-        });
-    let message = message.trim();
-    let (severity, message) = if let Some(message) = message.strip_prefix("warning:") {
-        (Severity::Warning, message.trim())
-    } else if let Some(message) = message.strip_prefix("error:") {
-        (Severity::Error, message.trim())
-    } else {
-        (Severity::Error, message)
-    };
-    if message.is_empty() {
-        return None;
-    }
-    Some(Diagnostic {
-        severity,
-        category: DiagnosticCategory::Compilation,
-        message: message.to_owned(),
-        location: Some(DiagnosticLocation {
-            path: compact_protobuf_path(path),
-            line: Some(line_number),
-            column,
-        }),
-        target: None,
-        action: None,
-        repetition_count: 1,
-    })
-}
-
-fn compact_protobuf_path(path: &str) -> String {
-    let path = strip_workspace_marker(path.trim_matches('"').replace('\\', "/"));
-    if let Some((_, after_execroot)) = path.rsplit_once("/execroot/")
-        && let Some((_, relative)) = after_execroot.split_once('/')
-    {
-        return relative.to_owned();
-    }
-    path.strip_prefix("./").unwrap_or(&path).to_owned()
-}
-
-fn parse_java_compiler_diagnostics(input: &str) -> Vec<Diagnostic> {
-    const MAX_CONTEXT_LINES: usize = 8;
-    let lines = input.lines().collect::<Vec<_>>();
-    let mut diagnostics = Vec::new();
-    for (index, line) in lines.iter().enumerate() {
-        let Some(mut diagnostic) = parse_java_compiler_diagnostic(line) else {
-            continue;
-        };
-        if diagnostic
-            .message
-            .eq_ignore_ascii_case("cannot find symbol")
-        {
-            for context in lines.iter().skip(index + 1).take(MAX_CONTEXT_LINES) {
-                if parse_java_compiler_diagnostic(context).is_some()
-                    || context.trim_start().starts_with("ERROR:")
-                {
-                    break;
-                }
-                if let Some(symbol) = context.trim().strip_prefix("symbol:") {
-                    diagnostic.message = format!(
-                        "cannot find symbol: {}",
-                        symbol.split_whitespace().collect::<Vec<_>>().join(" ")
-                    );
-                    break;
-                }
-            }
-        }
-        diagnostics.push(diagnostic);
-    }
-    diagnostics
-}
-
-fn parse_java_compiler_diagnostic(line: &str) -> Option<Diagnostic> {
-    let marker = line.rfind(".java:")?;
-    let path_end = marker + ".java".len();
-    let path = line[..path_end]
-        .trim()
-        .strip_prefix("ERROR: ")
-        .unwrap_or_else(|| line[..path_end].trim());
-    let (line_number, remainder) = split_u32_prefix(&line[path_end + 1..])?;
-    let (column, message) = split_u32_prefix(remainder)
-        .map_or((None, remainder), |(column, message)| {
-            (Some(column), message)
-        });
-    let message = message.trim();
-    let (severity, message) = if let Some(message) = message.strip_prefix("error:") {
-        (Severity::Error, message.trim())
-    } else {
-        let message = message.strip_prefix("warning:")?;
-        (Severity::Warning, message.trim())
-    };
-    if message.is_empty() {
-        return None;
-    }
-    Some(Diagnostic {
-        severity,
-        category: DiagnosticCategory::Compilation,
-        message: message.to_owned(),
-        location: Some(DiagnosticLocation {
-            path: compact_java_path(path),
-            line: Some(line_number),
-            column,
-        }),
-        target: None,
-        action: None,
-        repetition_count: 1,
-    })
-}
-
-/// Stateful extractor for Java exceptions followed by JVM stack frames.
-#[derive(Debug, Default)]
-pub struct JavaTestDiagnosticParser {
-    pending: Option<Diagnostic>,
-    pending_is_explicit: bool,
-    frames_seen: usize,
-}
-
-impl JavaTestDiagnosticParser {
-    const MAX_STACK_FRAMES: usize = 64;
-
-    /// Observes one normalized test-log line and emits an exception once an
-    /// application frame, another exception, or the end of its stack is seen.
-    pub fn observe_line(&mut self, line: &str) -> Option<Diagnostic> {
-        if let Some((message, explicitly_java)) = parse_java_exception_line(line) {
-            let previous = self.take_confirmed();
-            self.pending = Some(Diagnostic {
-                severity: Severity::Error,
-                category: DiagnosticCategory::Test,
-                message: message.to_owned(),
-                location: None,
-                target: None,
-                action: None,
-                repetition_count: 1,
-            });
-            self.pending_is_explicit = explicitly_java;
-            self.frames_seen = 0;
-            return previous;
-        }
-        self.pending.as_ref()?;
-        if let Some((location, framework_frame)) = parse_java_stack_frame(line) {
-            self.frames_seen = self.frames_seen.saturating_add(1);
-            if !framework_frame {
-                let mut diagnostic = self.pending.take()?;
-                diagnostic.location = Some(location);
-                self.pending_is_explicit = false;
-                self.frames_seen = 0;
-                return Some(diagnostic);
-            }
-            if self.frames_seen >= Self::MAX_STACK_FRAMES {
-                return self.take_confirmed();
-            }
-            return None;
-        }
-        if line.trim().is_empty() {
-            return None;
-        }
-        self.take_confirmed()
-    }
-
-    /// Emits an exception that reached end-of-file without an application frame.
-    pub fn finish(&mut self) -> Option<Diagnostic> {
-        self.take_confirmed()
-    }
-
-    fn take_confirmed(&mut self) -> Option<Diagnostic> {
-        let confirmed = self.pending_is_explicit || self.frames_seen > 0;
-        self.pending_is_explicit = false;
-        self.frames_seen = 0;
-        if confirmed {
-            self.pending.take()
-        } else {
-            self.pending = None;
-            None
-        }
-    }
-}
-
-fn java_exception_message(line: &str) -> Option<&str> {
-    parse_java_exception_line(line).map(|(message, _)| message)
-}
-
-fn parse_java_exception_line(line: &str) -> Option<(&str, bool)> {
-    let mut line = line.trim();
-    let mut explicitly_java = false;
-    if let Some(remainder) = line.strip_prefix("Exception in thread \"") {
-        let (_, remainder) = remainder.split_once("\" ")?;
-        line = remainder;
-        explicitly_java = true;
-    } else if let Some(remainder) = line.strip_prefix("Caused by: ") {
-        line = remainder;
-        explicitly_java = true;
-    }
-    let exception_type = line.split_once(':').map_or(line, |(name, _)| name);
-    if exception_type.is_empty()
-        || exception_type
-            .bytes()
-            .any(|byte| !(byte.is_ascii_alphanumeric() || matches!(byte, b'_' | b'.' | b'$')))
-    {
-        return None;
-    }
-    let class_name = exception_type.rsplit('.').next()?;
-    let recognized = class_name.ends_with("Error")
-        || class_name.ends_with("Exception")
-        || class_name.ends_with("Failure");
-    (recognized && (explicitly_java || exception_type.contains('.')))
-        .then_some((line, explicitly_java))
-}
-
-fn parse_java_stack_frame(line: &str) -> Option<(DiagnosticLocation, bool)> {
-    let frame = line.trim().strip_prefix("at ")?;
-    let (callable, source) = frame.split_once('(')?;
-    let source = source.strip_suffix(')')?;
-    let (file, line_number) = source.rsplit_once(':')?;
-    if !file.ends_with(".java") {
-        return None;
-    }
-    let line_number = line_number.parse::<u32>().ok()?;
-    let callable = callable.rsplit_once('/').map_or(callable, |(_, name)| name);
-    let class_name = callable.rsplit_once('.')?.0;
-    let package = class_name.rsplit_once('.').map(|(package, _)| package);
-    let path = package.map_or_else(
-        || file.to_owned(),
-        |package| format!("{}/{}", package.replace('.', "/"), file),
-    );
-    let framework_frame = [
-        "java.",
-        "javax.",
-        "jdk.",
-        "sun.",
-        "junit.",
-        "org.junit.",
-        "org.hamcrest.",
-        "org.opentest4j.",
-        "com.google.testing.junit.",
-    ]
-    .iter()
-    .any(|prefix| callable.starts_with(prefix));
-    Some((
-        DiagnosticLocation {
-            path,
-            line: Some(line_number),
-            column: None,
-        },
-        framework_frame,
-    ))
-}
-
-fn compact_java_path(path: &str) -> String {
-    let path = strip_workspace_marker(path.trim_matches('"').replace('\\', "/"));
-    if let Some((_, after_execroot)) = path.rsplit_once("/execroot/")
-        && let Some((_, relative)) = after_execroot.split_once('/')
-    {
-        return relative.to_owned();
-    }
-    path.strip_prefix("./").unwrap_or(&path).to_owned()
-}
-
-/// Stateful extractor for Bazel's Starlark source and traceback diagnostics.
-///
-/// Syntax diagnostics carry their location inline. Runtime Starlark failures
-/// instead print one or more `File "...", line N, column N` frames before a
-/// terminal `Error in ...` line, so only the latest frame must be retained.
-#[derive(Debug)]
-struct StarlarkDiagnosticParser {
-    location: Option<DiagnosticLocation>,
-    category: DiagnosticCategory,
-}
-
-impl Default for StarlarkDiagnosticParser {
-    fn default() -> Self {
-        Self {
-            location: None,
-            category: DiagnosticCategory::Loading,
-        }
-    }
-}
-
-impl StarlarkDiagnosticParser {
-    fn observe_line(&mut self, line: &str) -> Option<Diagnostic> {
-        if is_starlark_traceback_header(line) {
-            self.location = None;
-            if line.trim_start().starts_with("ERROR:") {
-                self.category = DiagnosticCategory::Loading;
-            }
-            return None;
-        }
-        if let Some(diagnostic) = parse_starlark_inline_diagnostic(line) {
-            self.location = diagnostic.location.clone();
-            self.category = diagnostic.category;
-            return is_starlark_root_cause_message(&diagnostic.message).then_some(diagnostic);
-        }
-        if let Some(location) = parse_starlark_traceback_location(line) {
-            self.location = Some(location);
-            return None;
-        }
-        let message = starlark_error_message(line)?;
-        Some(Diagnostic {
-            severity: Severity::Error,
-            category: self.category,
-            message: message.to_owned(),
-            location: self.location.take(),
-            target: None,
-            action: None,
-            repetition_count: 1,
-        })
-    }
-}
-
-fn parse_starlark_inline_diagnostic(line: &str) -> Option<Diagnostic> {
-    let line = line
-        .trim()
-        .strip_prefix("ERROR: ")
-        .unwrap_or_else(|| line.trim());
-    let path_end = starlark_path_end(line)?;
-    let path = &line[..path_end];
-    let (line_number, remainder) = split_u32_prefix(&line[path_end + 1..])?;
-    let (column, message) = split_u32_prefix(remainder)
-        .map_or((None, remainder), |(column, message)| {
-            (Some(column), message)
-        });
-    let message = message.trim();
-    if message.is_empty() {
-        return None;
-    }
-    let lower = message.to_ascii_lowercase();
-    let category = if (lower.starts_with("in ") && lower.contains(" rule //"))
-        || lower.contains("analysis of target")
-        || lower.contains("aspect on target")
-    {
-        DiagnosticCategory::Analysis
-    } else {
-        DiagnosticCategory::Loading
-    };
-    Some(Diagnostic {
-        severity: if lower.contains("warning:") {
-            Severity::Warning
-        } else {
-            Severity::Error
-        },
-        category,
-        message: message.to_owned(),
-        location: Some(DiagnosticLocation {
-            path: compact_starlark_path(path),
-            line: Some(line_number),
-            column,
-        }),
-        target: None,
-        action: None,
-        repetition_count: 1,
-    })
-}
-
-fn starlark_path_end(line: &str) -> Option<usize> {
-    const MARKERS: [&str; 6] = [
-        ".bzl:",
-        ".bazel:",
-        "/BUILD:",
-        "\\BUILD:",
-        "/WORKSPACE:",
-        "\\WORKSPACE:",
-    ];
-    MARKERS
-        .iter()
-        .filter_map(|marker| {
-            line.rfind(marker)
-                .map(|index| index + marker.len().saturating_sub(1))
-        })
-        .max()
-}
-
-fn parse_starlark_traceback_location(line: &str) -> Option<DiagnosticLocation> {
-    let marker = "File \"";
-    let start = line.find(marker)? + marker.len();
-    let remainder = &line[start..];
-    let (path, remainder) = remainder.split_once("\", line ")?;
-    if !is_starlark_path(path) {
-        return None;
-    }
-    let line_digits = remainder
-        .bytes()
-        .take_while(|byte| byte.is_ascii_digit())
-        .count();
-    if line_digits == 0 {
-        return None;
-    }
-    let line_number = remainder[..line_digits].parse::<u32>().ok()?;
-    let column = remainder[line_digits..]
-        .strip_prefix(", column ")
-        .and_then(|remainder| {
-            let digits = remainder
-                .bytes()
-                .take_while(|byte| byte.is_ascii_digit())
-                .count();
-            (digits > 0)
-                .then(|| remainder[..digits].parse::<u32>().ok())
-                .flatten()
-        });
-    Some(DiagnosticLocation {
-        path: compact_starlark_path(path),
-        line: Some(line_number),
-        column,
-    })
-}
-
-fn is_starlark_path(path: &str) -> bool {
-    path.ends_with(".bzl")
-        || path.ends_with(".bazel")
-        || matches!(path.rsplit(['/', '\\']).next(), Some("BUILD" | "WORKSPACE"))
-}
-
-fn is_starlark_traceback_header(line: &str) -> bool {
-    line.trim()
-        .strip_prefix("ERROR: ")
-        .unwrap_or_else(|| line.trim())
-        == "Traceback (most recent call last):"
-}
-
-fn starlark_error_message(line: &str) -> Option<&str> {
-    let line = line.trim();
-    (line.starts_with("Error in ") || line.starts_with("Error: ")).then_some(line)
-}
-
-fn is_starlark_root_cause_message(message: &str) -> bool {
-    let lower = message.to_ascii_lowercase();
-    lower.contains("syntax error")
-        || lower.contains("contains syntax errors")
-        || (lower.contains("name '") && lower.contains(" is not defined"))
-}
-
-fn compact_starlark_path(path: &str) -> String {
-    let path = strip_workspace_marker(path.trim_matches('"').replace('\\', "/"));
-    if let Some((_, after_execroot)) = path.rsplit_once("/execroot/")
-        && let Some((_, relative)) = after_execroot.split_once('/')
-    {
-        return relative.to_owned();
-    }
-    path.strip_prefix("./").unwrap_or(&path).to_owned()
-}
-
-/// Stateful extractor for standard Python traceback and syntax-error output.
-///
-/// Python reports source locations on a `File "...", line N` frame before the
-/// terminal exception. Keeping only the latest frame is bounded and matches
-/// traceback semantics, where the innermost frame is printed last.
-#[derive(Debug, Default)]
-pub struct PythonDiagnosticParser {
-    location: Option<DiagnosticLocation>,
-}
-
-impl PythonDiagnosticParser {
-    /// Observes one normalized output line and returns a diagnostic when the
-    /// line terminates a Python exception block.
-    pub fn observe_line(&mut self, line: &str) -> Option<Diagnostic> {
-        if line.trim() == "Traceback (most recent call last):" {
-            self.location = None;
-            return None;
-        }
-        if let Some(location) = parse_python_location(line) {
-            self.location = Some(location);
-            return None;
-        }
-        let message = python_exception_message(line)?;
-        let exception_type = message.split_once(':').map_or(message, |(name, _)| name);
-        Some(Diagnostic {
-            severity: if exception_type.ends_with("Warning") {
-                Severity::Warning
-            } else {
-                Severity::Error
-            },
-            category: DiagnosticCategory::Compilation,
-            message: message.to_owned(),
-            location: self.location.take(),
-            target: None,
-            action: None,
-            repetition_count: 1,
-        })
-    }
-}
-
-fn parse_python_location(line: &str) -> Option<DiagnosticLocation> {
-    let marker = "File \"";
-    let start = line.find(marker)? + marker.len();
-    let remainder = &line[start..];
-    let (path, remainder) = remainder.split_once("\", line ")?;
-    let synthetic_path = path.starts_with('<')
-        && !path.starts_with("<RUNTIME_ROOT>/")
-        && !path.starts_with("<WORKSPACE>/")
-        && !path.starts_with("<workspace>/");
-    if synthetic_path || path.ends_with("_stage2_bootstrap.py") {
-        return None;
-    }
-    let digits = remainder
-        .bytes()
-        .take_while(|byte| byte.is_ascii_digit())
-        .count();
-    if digits == 0 {
-        return None;
-    }
-    let line_number = remainder[..digits].parse::<u32>().ok()?;
-    Some(DiagnosticLocation {
-        path: compact_python_path(path),
-        line: Some(line_number),
-        column: None,
-    })
-}
-
-fn python_exception_message(line: &str) -> Option<&str> {
-    let mut line = line.trim();
-    if let Some(remainder) = line.strip_prefix('E')
-        && remainder.chars().next().is_some_and(char::is_whitespace)
-    {
-        line = remainder.trim_start();
-    }
-    if line.contains("File \"") {
-        return None;
-    }
-    let exception_type = line.split_once(':').map_or(line, |(name, _)| name);
-    if exception_type.is_empty()
-        || exception_type
-            .bytes()
-            .any(|byte| !(byte.is_ascii_alphanumeric() || matches!(byte, b'_' | b'.')))
-    {
-        return None;
-    }
-    let class_name = exception_type.rsplit('.').next()?;
-    let recognized = class_name.ends_with("Error")
-        || class_name.ends_with("Exception")
-        || class_name.ends_with("Failure")
-        || class_name.ends_with("Warning")
-        || matches!(class_name, "Failed" | "KeyboardInterrupt" | "SystemExit");
-    recognized.then_some(line)
-}
-
-fn compact_python_path(path: &str) -> String {
-    let path = strip_workspace_marker(path.trim_matches('"').replace('\\', "/"));
-    for marker in [".runfiles/_main/", ".runfiles/__main__/"] {
-        if let Some((_, relative)) = path.rsplit_once(marker) {
-            return relative.to_owned();
-        }
-    }
-    if let Some((_, after_execroot)) = path.rsplit_once("/execroot/")
-        && let Some((_, relative)) = after_execroot.split_once('/')
-    {
-        return relative.to_owned();
-    }
-    path.strip_prefix("./").unwrap_or(&path).to_owned()
-}
-
-/// Parses the standard Go compiler location form without depending on a
-/// particular diagnostic message or language setting.
-#[must_use]
-pub fn parse_go_diagnostic(line: &str) -> Option<Diagnostic> {
-    let marker = line.rfind(".go:")?;
-    let path_end = marker + ".go".len();
-    let path = line[..path_end]
-        .trim()
-        .strip_prefix("ERROR: ")
-        .unwrap_or_else(|| line[..path_end].trim());
-    let (line_number, remainder) = split_u32_prefix(&line[path_end + 1..])?;
-    let (column, message) = split_u32_prefix(remainder)
-        .map_or((None, remainder), |(column, message)| {
-            (Some(column), message)
-        });
-    let message = message.trim();
-    if message.is_empty() {
-        return None;
-    }
-    Some(Diagnostic {
-        severity: if message.to_ascii_lowercase().contains("warning:") {
-            Severity::Warning
-        } else {
-            Severity::Error
-        },
-        category: DiagnosticCategory::Compilation,
-        message: message.to_owned(),
-        location: Some(DiagnosticLocation {
-            path: compact_go_path(path),
-            line: Some(line_number),
-            column,
-        }),
-        target: None,
-        action: None,
-        repetition_count: 1,
-    })
-}
-
-fn split_u32_prefix(value: &str) -> Option<(u32, &str)> {
-    let (number, remainder) = value.split_once(':')?;
-    let number = number.trim();
-    (!number.is_empty() && number.bytes().all(|byte| byte.is_ascii_digit()))
-        .then(|| number.parse::<u32>().ok().map(|number| (number, remainder)))
-        .flatten()
-}
-
-fn strict_dependency_diagnostic(line: &str) -> Option<Diagnostic> {
-    const MARKER: &str = ": import of \"";
-    let marker = line.find(MARKER)?;
-    let path = line[..marker].trim();
-    if !path.ends_with(".go") {
-        return None;
-    }
-    let import = line[marker + MARKER.len()..].split('"').next()?.trim();
-    if import.is_empty() {
-        return None;
-    }
-    let path = compact_go_path(path);
-    Some(Diagnostic {
-        severity: Severity::Error,
-        category: DiagnosticCategory::Compilation,
-        message: format!(
-            "missing strict dependency: {path} imports \"{import}\"; add its target to deps"
-        ),
-        location: Some(DiagnosticLocation {
-            path,
-            line: None,
-            column: None,
-        }),
-        target: None,
-        action: None,
-        repetition_count: 1,
-    })
-}
-
-fn compact_go_path(path: &str) -> String {
-    let path = strip_workspace_marker(path.trim_matches('"').replace('\\', "/"));
-    if let Some((_, after_execroot)) = path.rsplit_once("/execroot/")
-        && let Some((_, relative)) = after_execroot.split_once('/')
-    {
-        return relative.to_owned();
-    }
-    path
-}
-
-fn strip_workspace_marker(path: String) -> String {
-    path.strip_prefix("<WORKSPACE>/")
-        .or_else(|| path.strip_prefix("<workspace>/"))
-        .unwrap_or(&path)
-        .to_owned()
-}
-
-fn is_actionable(line: &str) -> bool {
-    let line = line.trim();
-    let lower = line.to_ascii_lowercase();
-    if matches!(lower.as_str(), "failure:" | "failures:")
-        || (line.starts_with("test ") && lower.ends_with(" ... ok"))
-    {
-        return false;
-    }
-    lower.contains("error:")
-        || lower.starts_with("error ")
-        || lower.contains("failed:")
-        || lower.contains("no such target")
-        || lower.contains("no such package")
-        || lower.contains("visibility error")
-        || lower.contains("undefined reference")
-        || lower.contains("fatal:")
-        || lower.contains("root_cause")
-        || lower.contains("panicked at")
-        || (lower.contains("assertion") && lower.contains(" failed"))
-        || lower.starts_with("test result: failed")
-        || (line.starts_with("test ") && line.ends_with(" ... FAILED"))
-}
-
-fn category_from_text(line: &str) -> DiagnosticCategory {
-    let lower = line.to_ascii_lowercase();
-    if lower.contains("no such package") || lower.contains("no such target") {
-        DiagnosticCategory::Loading
-    } else if lower.contains("visibility") {
-        DiagnosticCategory::Visibility
-    } else if lower.contains("analysis") {
-        DiagnosticCategory::Analysis
-    } else if lower.contains("test") || lower.contains("panicked at") || lower.contains("assertion")
-    {
-        DiagnosticCategory::Test
-    } else if lower.contains("error:")
-        || lower.contains("error[")
-        || lower.contains("undefined reference")
-    {
-        DiagnosticCategory::Compilation
-    } else if lower.contains("root_cause") {
-        DiagnosticCategory::Test
-    } else {
-        DiagnosticCategory::Unknown
-    }
-}
-
 fn abort_category(reason: i32) -> DiagnosticCategory {
     match reason {
         5 => DiagnosticCategory::Loading,
@@ -2444,6 +890,11 @@ fn test_status(value: i32) -> TestStatus {
 
 #[cfg(test)]
 mod tests {
+    use crate::diagnostics::{
+        PythonDiagnosticParser, cpp_path_end, parse_cpp_diagnostic, parse_cpp_linker_diagnostic,
+        parse_protobuf_diagnostic, parse_starlark_inline_diagnostic, parse_swc_diagnostics,
+        parse_typescript_diagnostic,
+    };
     use bazel_mcp_bep::proto::{
         ActionExecuted, BuildEvent, BuildEventId, File, FileOwnedView, NamedSetOfFiles,
     };
@@ -2452,6 +903,7 @@ mod tests {
         file as owned_file,
     };
     use bazel_mcp_bep::{BepEvent, encode_event_id};
+    use bazel_mcp_types::DiagnosticLocation;
 
     use super::*;
